+++ src/test/java/io/github/astrapi69/lang/model/AnnotationModelTest.java
@@ -20,10 +20,10 @@
  */
 package io.github.astrapi69.lang.model;
 
-import static org.testng.AssertJUnit.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
+import org.junit.jupiter.api.Test;
 import org.meanbean.test.BeanTester;
-import org.testng.annotations.Test;
 
 import nl.jqno.equalsverifier.EqualsVerifier;