+++ src/test/java/io/github/astrapi69/lang/model/ClassModelTest.java
@@ -20,7 +20,7 @@
  */
 package io.github.astrapi69.lang.model;
 
-import static org.testng.AssertJUnit.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 import java.util.List;
 import java.util.Map;
@@ -28,8 +28,8 @@ import java.util.Map;
 import javax.lang.model.element.ElementKind;
 import javax.lang.model.element.Modifier;
 
+import org.junit.jupiter.api.Test;
 import org.meanbean.test.BeanTester;
-import org.testng.annotations.Test;
 
 import io.github.astrapi69.collection.list.ListFactory;
 import io.github.astrapi69.collection.map.MapFactory;
@@ -56,7 +56,8 @@ public class ClassModelTest
 
 		/** The imports. */
 		List<String> imports = ListFactory.newArrayList(
-			"import static org.testng.AssertJUnit.assertNotNull;", "import java.util.Map;");
+			"import static org.junit.jupiter.api.Assertions.assertNotNull;",
+			"import java.util.Map;");
 
 		/** The class annotations. */
 		List<ClassModel> classAnnotations = ListFactory.newArrayList();